same wordtype only

+++ app/models/term.rb
@@ -7,11 +7,14 @@ class Term < ActiveRecord::Base
     @rez = [] 
     if term = Term.find_by_word(word)
       Termlink.where("term_id = ?", term.id).each do |link|
-        rijec = Term.find(link.link_id)
-        if @rez.map{|a| a.word}.include? rijec.word
-          link.destroy
-        else
-          @rez.push rijec
+        rijec = Term.where(:id => link.link_id,
+        :wordtype => term.wordtype).first
+        unless rijec.nil?    
+          if @rez.map{|a| a.word}.include? rijec.word
+            link.destroy
+          else          
+            @rez.push rijec
+          end
         end
       end
      end
